Use users auth control file path as user key
Since openvpn 2.4 ip and port for user can float, this causes ghost
sessions to appear, also, part of accounting data is lost. So key
format IP:PORT is no longer valid. User control file is uniqe per
session, so it seems to be ok to use. Ideally, we should use session
id if we can get one in plugin.